+++ packages/react-native-sdk/src/components/Participant/FloatingParticipantView/index.tsx
@@ -66,6 +66,7 @@ const CustomLocalParticipantViewVideoFallback = () => {
       colors,
       floatingParticipantsView,
       variants: { iconSizes },
+      defaults,
     },
   } = useTheme();
 
@@ -73,12 +74,12 @@ const CustomLocalParticipantViewVideoFallback = () => {
     <View
       style={[
         styles.videoFallback,
-        { backgroundColor: colors.disabled },
+        { backgroundColor: colors.sheetSecondary },
         floatingParticipantsView.videoFallback,
       ]}
     >
       <View style={{ height: iconSizes.md, width: iconSizes.md }}>
-        <VideoSlash color={colors.static_white} />
+        <VideoSlash color={colors.iconPrimary} size={defaults.iconSize} />
       </View>
     </View>
   );
@@ -170,9 +171,7 @@ export const FloatingParticipantView = ({
                 style={[
                   styles.participantViewContainer,
                   participantViewStyle,
-                  {
-                    shadowColor: colors.static_black,
-                  },
+                  { shadowColor: colors.sheetPrimary },
                   floatingParticipantsView.participantViewContainer,
                 ]}
                 // video z order must be one above the one used in grid view
@@ -192,7 +191,6 @@ export const FloatingParticipantView = ({
 
 const styles = StyleSheet.create({
   container: {
-    margin: 8,
     // Needed to make the view on top and draggable
     zIndex: Z_INDEX.IN_MIDDLE,
     flex: 1,