+++ sample-apps/react-native/dogfood/src/components/LobbyViewComponent.tsx
@@ -31,18 +31,7 @@ export const LobbyViewComponent = ({
     return (
       <>
         <JoinCallButton onPressHandler={onJoinCallHandler} />
-        {route.name === 'MeetingScreen' ? (
-          <Pressable
-            style={styles.anonymousButton}
-            onPress={() => {
-              navigation.navigate('GuestModeScreen', { callId });
-            }}
-          >
-            <Text style={styles.anonymousButtonText}>
-              {t('Join as Guest or Anonymously')}
-            </Text>
-          </Pressable>
-        ) : (
+        {route.name !== 'MeetingScreen' && (
           <Pressable
             style={styles.anonymousButton}
             onPress={() => {